Refactor useToast and useModal
