+++ packages/ui/src/components/va-modal/index.ts
@@ -2,3 +2,4 @@ import withConfigTransport from '../../services/config-transport/withConfigTrans
 import _VaModal from './VaModal.vue'
 
 export const VaModal = withConfigTransport(_VaModal)
+export { useModal } from './hooks/useModal'
