+++ packages/ui/src/components/va-toast/index.ts
@@ -1,0 +1,6 @@
+import withConfigTransport from '../../services/config-transport/withConfigTransport'
+import _VaToast from './VaToast.vue'
+
+export const VaToast = withConfigTransport(_VaToast)
+
+export { useToast } from './hooks/useToast'
